renamed quote section files with jsx extension and reformatted code
