+++ app/components/index.js
@@ -3,9 +3,4 @@ import InputText from './InputText.jsx';
 import Container from './Container.jsx';
 import VideoPlayer from './VideoPlayer.jsx';
 
-export {
-    TextSubmit,
-    InputText,
-    Container,
-    VideoPlayer
-};
+export { TextSubmit, InputText, Container, VideoPlayer };
