--- app/pages/blackout/QuoteSection/company_name.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export default styled.div`
-  color: ${props => props.dark ? props.theme.darkPink : props.theme.pink};
-  fontStyle: italic;
-  textAlign: ${props => (props.position === 'right') ? 'right' : 'left'};
-`;
